pyproject.toml
[tool.poetry]
name = "github-yaml-modifier"
version = "0.1.0"
description = ""
authors = ["heumsi <heumsi@naver.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
python-dotenv = "^0.17.1"
glom = "^20.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
